Add backchannel NSS provider query on initgr calls
This is needed in order to assure the memcache is properly and promptly
cleaned up if a user memberships change on login.

The list of the current groups for the user is sourced before it is
updated and sent to the NSS provider to verify if it has changed after
the update call has been made.