select_principal_from_keytab() do wildcard lookups after specific ones
Currently the wildcard lookup '*$' is done before the one for
host/our.hostname@REALM. This means we would ignore a more specific
match in favour of an unspecific match with a principal which is only
used in a AD environment.

I think this is wrong an wildcards should only be used is all specific
lookups fail.