set ret to EOK after for loop in sdap_sudo_purge_sudoers
If we are unable to delete some rule from cache we print a debug
message and ignore the error. Thus we should set ret to EOK after
the for loop otherwise we return whether the last rule was deleted
successfully or not.

This also removes compilation warning that ret may be used
uninitialized (when we don't go inside the loop at all).