Remove unhelpful vtable from sss_cache
Using a vtable like this has various drawacks, including the fact prototypes
are not checked by the compiler so the code could silently break and still
compile fine (in fact I found this out changing one of the prototypes).

A switch statement is also better because it catches if the enum changed and
won't risk allowing to access the table out of bounds.