Invalidate user entry even if there are no groups
Related to https://fedorahosted.org/sssd/ticket/1757

Previously we would optimize the mc invalidate code for cases where the
user was a member of some groups. But if the user was removed from the
server while being in memory cache, we would only invalidate the mc
record if he was a member of at least one supplementary group.