Provide libnl3 support
https://fedorahosted.org/sssd/ticket/812

Update the monitor code to be using the new libnl3 API.

Changed configure option
--with-libnl

By default, it tries to build with libnl3, if not found, then with
libnl1, if this isn't found either, build proceeds without libnl, just
with warning.

Specifing --with-libnl=<libnl3|libnl1|no> checks for the specific given
version, if not found, configure ends with error.