BUILD: Fix cmocka detection
We were not properly detecting that cmocka was unavailable. It was
expecting an empty value and getting "no" instead. This patch
corrects the expectation, so we will now skip building and running
cmocka tests on platforms that do not have it available.

Also, we were missing the cmocka header files in the distribution
tarball, so 'make distcheck' was failing.