Default TEST_DIR to cwd, not empty string if not set explicitly
If configure isn't being run with argument --with-test-dir, then variable
TEST_DIR will be defined, but its value will be empty (""). In this case
opendir will fail with uncatched error "Directory does not exist, or name is
an empty string". Finally function call dirfd will segfault because its
argument is NULL.

I changed default value of TEST_DIR (if --with-test-dir was not used).

Function tests_set_cwd does not ignore return value of chdir, because
TEST_DIR should not be the empty string.