Always update cached upn if enterprise principals are used
Instead of continuing to use the initial upn if enterprise principals
are used if should always be replaced. The enterprise principal
is stored in the credential cache and without knowing it the
ccache_for_princ() calls to determine the location of the credential
cache will fail.

Fixes https://fedorahosted.org/sssd/ticket/1921