Intermittent fix for get_user_and_group_users_done
users_get_recv() never returns ENOENT. In general it should return EOK
in the case no matching user was found. But since I forget to handle a
SID based filter properly in sdap_get_users_process() an error is
returned in this case which makes get_user_and_group_users_done() work
as expected with this patch.

There is an upcoming enhancement to users_get_recv() which I'm planning
to use for a full fix.