idmap: add sss_idmap_domain_has_algorithmic_mapping
With this call it can be checked if for a given domain algorithmic
mapping is available or if the ID must be read from an external source.
The default if an error occurs or no matching range was found is false,
i.e external mapping, to meet the requirements for simple LDAP based
domains where only external mapping is available.

Fixes https://fedorahosted.org/sssd/ticket/1960