failover: return error when SRV lookup returned only duplicates
https://fedorahosted.org/sssd/ticket/1947

Otherwise we risk that the meta server is removed from the server list,
but without a chance to return, because there may be no fo_server with
srv_data = meta.

Also if state->meta->next is NULL (it is still orphaned because we try
to errornously expand it without invoking collapse first), state->out
will be NULL and SSSD will crash.

New error code: ERR_SRV_DUPLICATES