AD: kinit with the local DC even when talking to a GC
We tried to use the GC address even for kinit which gave us errors like:
"Realm not local to KDC while getting initial credentials".

This patch adds a new AD_GC service that is only used for ID lookups,
any sort of Kerberos operations are done against the local servers.